refactor(agent_supervisor): extract spawn_subordinate (#1266)
lifecycle.rs 467 -> lifecycle/{mod.rs ~303, spawn.rs ~179}.
